Resolve conflicts:
- `Settings.cpp`: keep both PR's `use_sparse_lightweight_representation_of_primary_key_for_index_analysis` and master's `use_partition_pruning`
- `IFunctionCustomWeek.h`: use master's `is_not_monotonic` variable pattern and structure
- `gtest_conversion_monotonic.cpp`: combine includes from both branches
- `KeyCondition.cpp`: keep PR's `<limits>` and `<Common/PODArray.h>` includes
- Test 03789: use master's more comprehensive version with DateTime, LowCardinality, Nullable tests

+++ src/Storages/MergeTree/DataPartsExchange.cpp
@@ -95,7 +95,7 @@ struct ReplicatedFetchReadCallback
         if (replicated_fetch_entry->total_size_bytes_compressed != 0)
         {
             replicated_fetch_entry->progress.store(
-                    static_cast<double>(bytes_count) / replicated_fetch_entry->total_size_bytes_compressed,
+                    static_cast<double>(bytes_count) / static_cast<double>(replicated_fetch_entry->total_size_bytes_compressed),
                     std::memory_order_relaxed);
         }
     }
